#27 refactor existing Jetty VaadinBoot on top of common VaadinBootBase
