+++ vaadin-boot/build.gradle.kts
@@ -3,20 +3,14 @@ plugins {
 }
 
 dependencies {
+    api(project(":common"))
     implementation(libs.slf4j.api)
-    api(libs.jetbrains.annotations)
 
     // Embedded Jetty dependencies.
     // This one is needed to host webapps and perform classpath scanning for annotations
     api(libs.jetty.webapp)
     // This one is required to have websocket/push support.
     implementation(libs.jetty.websocket)
-
-    // opens url in a browser
-    implementation(libs.vaadin.open)
-
-    // both Vaadin and vaadin-open depends on commons-io, we can use it too.
-    implementation(libs.commons.io)
 
     testImplementation(libs.slf4j.simple)
     testImplementation(libs.junit)
