+++ vaadin-boot/src/main/java/com/github/mvysny/vaadinboot/VaadinBoot.java
@@ -1,27 +1,12 @@
 package com.github.mvysny.vaadinboot;
 
-import com.vaadin.open.Open;
-import jakarta.servlet.Servlet;
-import org.eclipse.jetty.ee10.webapp.MetaInfConfiguration;
+import com.github.mvysny.vaadinboot.common.JettyWebServer;
+import com.github.mvysny.vaadinboot.common.VaadinBootBase;
+import com.github.mvysny.vaadinboot.common.WebServer;
 import org.eclipse.jetty.ee10.webapp.WebAppContext;
-import org.eclipse.jetty.server.Server;
-import org.eclipse.jetty.server.ServerConnector;
-import org.eclipse.jetty.util.resource.Resource;
-import org.eclipse.jetty.util.thread.QueuedThreadPool;
-import org.eclipse.jetty.util.thread.ThreadPool;
 import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
-import org.jetbrains.annotations.VisibleForTesting;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
-import java.lang.reflect.Method;
-import java.net.InetSocketAddress;
-import java.time.Duration;
-import java.util.Objects;
-import java.util.concurrent.Executor;
-import java.util.concurrent.Executors;
 
 /**
  * Bootstraps your Vaadin application from your main() function. Simply call
@@ -33,45 +18,7 @@ import java.util.concurrent.Executors;
  * By default, listens on all interfaces; call {@link #localhostOnly()} to only
  * listen on localhost.
  */
-public class VaadinBoot {
-    /**
-     * The default port where Jetty will listen for http:// traffic.
-     */
-    private static final int DEFAULT_PORT = 8080;
-
-    /**
-     * The port where Jetty will listen for http:// traffic. Defaults to {@value #DEFAULT_PORT}.
-     * <br/>
-     * Can be configured via the <code>SERVER_PORT</code> environment variable, or <code>-Dserver.port=</code> Java system property.
-     */
-    @VisibleForTesting
-    int port = Integer.parseInt(Env.getProperty("SERVER_PORT", "server.port", "" + DEFAULT_PORT));
-
-    /**
-     * Listen on interface handling given host name. Defaults to <code>null</code> which causes Jetty
-     * to listen on all interfaces.
-     * <br/>
-     * Can be configured via the <code>SERVER_ADDRESS</code> environment variable, or <code>-Dserver.address=</code> Java system property.
-     */
-    @Nullable
-    @VisibleForTesting
-    String hostName = Env.getProperty("SERVER_ADDRESS", "server.address");
-
-    /**
-     * The context root to run under. Defaults to `/`.
-     * Change this to e.g. /foo to host your app on a different context root
-     * <br/>
-     * Can be configured via the <code>SERVER_SERVLET_CONTEXT-PATH</code> environment variable, or <code>-Dserver.servlet.context-path=</code> Java system property.
-     */
-    @NotNull
-    @VisibleForTesting
-    String contextRoot = Env.getProperty("SERVER_SERVLET_CONTEXT-PATH", "server.servlet.context-path", "/");
-
-    /**
-     * When the app launches, open the browser automatically when in dev mode.
-     */
-    private boolean openBrowserInDevMode = true;
-
+public class VaadinBoot extends VaadinBootBase<VaadinBoot> {
     /**
      * If true, no classpath scanning is performed - no servlets nor weblisteners are detected.
      * <br/>
@@ -103,80 +50,7 @@ public class VaadinBoot {
      * Creates new boot instance.
      */
     public VaadinBoot() {
-    }
-
-    /**
-     * Sets the port to listen on. Listens on {@value #DEFAULT_PORT} by default.
-     * @param port the new port, 1..65535
-     * @return this
-     */
-    @NotNull
-    public VaadinBoot setPort(int port) {
-        if (port < 1 || port > 65535) {
-            throw new IllegalArgumentException("Parameter port: invalid value " + port + ": must be 1..65535");
-        }
-        this.port = port;
-        return this;
-    }
-
-    /**
-     * Listen on network interface handling given host name. Pass in null to listen on all interfaces;
-     * pass in `127.0.0.1` or `localhost` to listen on localhost only (or call {@link #localhostOnly()}).
-     * @param hostName the interface to listen on.
-     * @return this
-     */
-    @NotNull
-    public VaadinBoot listenOn(@Nullable String hostName) {
-        this.hostName = hostName;
-        return this;
-    }
-
-    /**
-     * Listen on the <code>localhost</code> network interface only.
-     * @return this
-     */
-    @NotNull
-    public VaadinBoot localhostOnly() {
-        return listenOn("localhost");
-    }
-
-    /**
-     * Change this to e.g. /foo to host your app on a different context root
-     * @param contextRoot the new context root, e.g. `/foo`.
-     * @return this
-     */
-    @NotNull
-    public VaadinBoot withContextRoot(@NotNull String contextRoot) {
-        this.contextRoot = Objects.requireNonNull(contextRoot);
-        return this;
-    }
-
-    /**
-     * When the app launches, open the browser automatically when in dev mode.
-     * @param openBrowserInDevMode defaults to true.
-     * @return this
-     */
-    @NotNull
-    public VaadinBoot openBrowserInDevMode(boolean openBrowserInDevMode) {
-        this.openBrowserInDevMode = openBrowserInDevMode;
-        return this;
-    }
-
-    /**
-     * When the app launches, open the browser automatically when in dev mode.
-     * @return if true, open the browser automatically when in dev mode.
-     */
-    public boolean isOpenBrowserInDevMode() {
-        return openBrowserInDevMode;
-    }
-
-    /**
-     * Returns the URL where the app is running, for example <code>http://localhost:8080/app</code>.
-     * @return the server URL, not null.
-     */
-    @NotNull
-    public String getServerURL() {
-        return "http://" + (hostName != null ? hostName : "localhost") + ":" + port + contextRoot;
+        super(new JettyWebServer());
     }
 
     /**
@@ -262,97 +136,9 @@ public class VaadinBoot {
         return useVirtualThreadsIfAvailable;
     }
 
-    // mark volatile: might be accessed by the shutdown hook from a different thread.
-    private volatile Server server;
-
-    /**
-     * Runs your app. Blocks until the user presses Enter or CTRL+C.
-     * <br/>
-     * WARNING: this function may never terminate when the entire JVM may be killed on CTRL+C.
-     * <br/>
-     * If you wish to wait for the Enter press yourself (if you want to control the lifecycle yourself),
-     * call {@link #start()}/{@link #stop(String)} instead of this function.
-     * @throws Exception when the webapp fails to start.
-     */
-    public void run() throws Exception {
-        start();
-
-        // We want to shut down the app cleanly by calling stop().
-        // Unfortunately, that's not easy. When running from:
-        // * Intellij as a Java app: CTRL+C doesn't work but Enter does.
-        // * ./gradlew run: Enter doesn't work (no stdin); CTRL+C kills the app forcibly.
-        // * ./mvnw exec:java: both CTRL+C and Enter works properly.
-        // * cmdline as an unzipped app (production): both CTRL+C and Enter works properly.
-        // Therefore, we'll use a combination of the two.
-
-        // this gets called both when CTRL+C is pressed, and when main() terminates.
-        Runtime.getRuntime().addShutdownHook(new Thread(() -> stop("Shutdown hook called, shutting down")));
-        System.out.println("Press ENTER or CTRL+C to shutdown");
-
-        if (isOpenBrowserInDevMode() && !Env.isVaadinProductionMode) {
-            Open.open(getServerURL());
-        }
-
-        // Await for Enter.
-        if (System.in.read() == -1) {
-            // "./gradlew" run offers no stdin and read() will return immediately with -1
-            // This happens when we're running from Gradle; but also when running from Docker with no tty
-            System.out.println("No stdin available. press CTRL+C to shutdown");
-            server.join(); // blocks endlessly
-        } else {
-            // Enter pressed - shut down.
-            stop("Main: Shutting down");
-        }
-    }
-
-    /**
-     * Starts the Jetty server and your app. Blocks until the app is fully started, then returns.
-     * Mostly used for testing.
-     * @throws Exception when the webapp fails to start.
-     */
-    public void start() throws Exception {
-        final long startupMeasurementSince = System.currentTimeMillis();
-        log.info("Starting App");
-
-        // detect&enable production mode, but only if it hasn't been specified by the user already
-        if (System.getProperty("vaadin.productionMode") == null && Env.isVaadinProductionMode) {
-            // fixes https://github.com/mvysny/vaadin14-embedded-jetty/issues/1
-            System.setProperty("vaadin.productionMode", "true");
-        }
-
-        fixClasspath();
-        log.debug("Classpath fixed");
-
-        final WebAppContext context = createWebAppContext();
-        log.debug("Jetty WebAppContext created");
-
-        server = new Server(newThreadPool());
-        final ServerConnector serverConnector = new ServerConnector(server);
-        serverConnector.setPort(port);
-        if (hostName != null) {
-            serverConnector.setHost(hostName);
-        }
-        server.addConnector(serverConnector);
-        server.setHandler(context);
-        log.debug("Jetty Server configured");
-        try {
-            server.start();
-            log.debug("Jetty Server started");
-
-            onStarted(context);
-
-            final Duration startupDuration = Duration.ofMillis(System.currentTimeMillis() - startupMeasurementSince);
-            System.out.println("\n\n=================================================\n" +
-                    "Started in " + startupDuration + ". Running on " + Env.dumpHost() + "\n" +
-                    "Please open " + getServerURL() + " in your browser.");
-            if (!Env.isVaadinProductionMode) {
-                System.out.println("If you see the 'Unable to determine mode of operation' exception, just kill me and run `./gradlew vaadinPrepareFrontend` or `./mvnw vaadin:prepare-frontend`");
-            }
-            System.out.println("=================================================\n");
-        } catch (Exception e) {
-            stop("Failed to start");
-            throw e;
-        }
+    @Override
+    protected void onStarted(@NotNull WebServer server) throws IOException {
+        onStarted(((JettyWebServer) server).getContext());
     }
 
     /**
@@ -362,92 +148,5 @@ public class VaadinBoot {
      * @throws IOException on i/o exception
      */
     protected void onStarted(@NotNull WebAppContext context) throws IOException {
-    }
-
-    /**
-     * Creates the Jetty {@link WebAppContext}.
-     * @return the {@link WebAppContext}
-     * @throws IOException on i/o exception
-     */
-    @NotNull
-    protected WebAppContext createWebAppContext() throws IOException {
-        final WebAppContext context = new WebAppContext();
-        final Resource webRoot = Env.findWebRoot(context.getResourceFactory());
-        context.setBaseResource(webRoot);
-        context.setContextPath(contextRoot);
-
-        // don't add the servlet this way - the @WebServlet annotation is ignored!
-        // https://github.com/mvysny/vaadin-boot/issues/22
-//        context.addServlet(servlet, "/*");
-
-        // when the webapp fails to initialize, make sure that start() throws.
-        context.setThrowUnavailableOnStartupException(true);
-        if (!isDisableClasspathScanning()) {
-            // this will properly scan the classpath for all @WebListeners, including the most important
-            // com.vaadin.flow.server.startup.ServletContextListeners.
-            // See also https://mvysny.github.io/vaadin-lookup-vs-instantiator/
-            // Jetty documentation: https://www.eclipse.org/jetty/documentation/jetty-12/operations-guide/index.html#og-annotations-scanning
-            String pattern = ".*\\.jar|.*/classes/.*";
-            if (isScanTestClasspath()) {
-                pattern += "|.*/test-classes/.*";
-            }
-            context.setAttribute(MetaInfConfiguration.CONTAINER_JAR_PATTERN, pattern);
-            // must be set to true, to enable classpath scanning:
-            // https://eclipse.dev/jetty/documentation/jetty-12/operations-guide/index.html#og-annotations-scanning
-            context.setConfigurationDiscovered(true);
-        }
-        return context;
-    }
-
-    /**
-     * See {@link Env#fixClasspath()}.
-     */
-    protected void fixClasspath() {
-        Env.fixClasspath();
-    }
-
-    /**
-     * Stops your app. Blocks until the webapp is fully stopped. Mostly used for tests.
-     * Never throws an exception.
-     * @param reason why we're shutting down. Logged as info.
-     */
-    public void stop(@NotNull String reason) {
-        try {
-            if (server != null) {
-                log.info(reason);
-                server.stop(); // blocks until the webapp stops fully
-                log.info("Stopped");
-                server = null;
-            }
-        } catch (Throwable t) {
-            log.error("stop() failed: " + t, t);
-        }
-    }
-
-    @NotNull
-    private static final Logger log = LoggerFactory.getLogger(VaadinBoot.class);
-
-    /**
-     * Creates a thread pool for Jetty to serve http requests.
-     * @return the thread pool, may be null if the default one is to be used.
-     */
-    @Nullable
-    protected ThreadPool newThreadPool() {
-        if (isUseVirtualThreadsIfAvailable() && Env.getJavaVersion() >= 21) {
-            log.info("Configuring Jetty to use JVM 21+ virtual threads");
-            // see https://eclipse.dev/jetty/documentation/jetty-12/programming-guide/index.html#pg-arch-threads-thread-pool-virtual-threads
-            final QueuedThreadPool threadPool = new QueuedThreadPool();
-            try {
-                // reflection: we call Java 21+ method, however we're compiled with Java 17
-                final Method m = Executors.class.getDeclaredMethod("newVirtualThreadPerTaskExecutor");
-                threadPool.setVirtualThreadsExecutor(((Executor) m.invoke(null)));
-            } catch (Exception ex) {
-                throw new RuntimeException(ex);
-            }
-            return threadPool;
-        } else {
-            log.info("Configuring Jetty to use regular JVM threads");
-            return null;
-        }
     }
 }
